Enable Kernel Samepage Merging (KSM)
Because the Samsung Galaxy Young has a limited amount of RAM,
we enable "Kernel Samepage Merging" (CONFIG_KSM) in royss_custom_defconfig.

KSM is a kernel thread that runs in the background and compares pages in memory that have been marked MADV_MERGEABLE by user-space. If two pages are found to be the same, the KSM thread merges them back as a single copy-on-write page of memory.

KSM will save memory over time on a running system, gaining memory duplication at a cost of CPU power, which could have an impact on battery life. You should measure whether the power tradeoff is worth the memory savings you get by enabling KSM.